git-update/rootfs/app/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.1
gitpython==3.1.42
watchdog==4.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
httpx==0.27.0
paho-mqtt==1.6.1
python-dotenv==1.0.1
rich==13.7.1
